Move related methods and variables in GameViewModel
