+++ app/src/main/java/com/example/android/guesstheword/screens/game/GameViewModel.kt
@@ -4,12 +4,71 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 
 class GameViewModel:ViewModel() {
+    // The current word
+    private var word = ""
+
+    // The current score
+    private var score = 0
+
+    // The list of words - the front of the list is the next word to guess
+    private lateinit var wordList: MutableList<String>
+
     init {
         Log.i("GameViewModel", "GameViewModel created")
+        resetList()
+        nextWord()
     }
 
     override fun onCleared() {
         super.onCleared()
         Log.i("GameViewModel", "GameViewModel destroyed")
+    }
+
+    private fun resetList() {
+        wordList = mutableListOf(
+                "queen",
+                "hospital",
+                "basketball",
+                "cat",
+                "change",
+                "snail",
+                "soup",
+                "calendar",
+                "sad",
+                "desk",
+                "guitar",
+                "home",
+                "railway",
+                "zebra",
+                "jelly",
+                "car",
+                "crow",
+                "trade",
+                "bag",
+                "roll",
+                "bubble"
+        )
+        wordList.shuffle()
+    }
+
+    private fun nextWord() {
+        //Select and remove a word from the list
+        if (wordList.isEmpty()) {
+            gameFinished()
+        } else {
+            word = wordList.removeAt(0)
+        }
+        updateWordText()
+        updateScoreText()
+    }
+
+    private fun onSkip() {
+        score--
+        nextWord()
+    }
+
+    private fun onCorrect() {
+        score++
+        nextWord()
     }
 }
